+++ lib/src/service/twitter_service.dart
@@ -12,6 +12,7 @@ import 'geo/geo_service.dart';
 import 'lists/lists_service.dart';
 import 'media/media_service.dart';
 import 'spaces/spaces_service.dart';
+import 'trends/trends_service.dart';
 import 'tweets/tweets_service.dart';
 import 'users/users_service.dart';
 
@@ -41,6 +42,9 @@ abstract class TwitterService {
   /// Returns the geo service.
   GeoService get geoService;
 
+  /// Returns the trends service.
+  TrendsService get trendsService;
+
   /// Returns the compliance service.
   ComplianceService get complianceService;
 }
@@ -55,6 +59,7 @@ class _TwitterService implements TwitterService {
         mediaService = MediaService(context: context),
         directMessagesService = DirectMessagesService(context: context),
         geoService = GeoService(context: context),
+        trendsService = TrendsService(context: context),
         complianceService = ComplianceService(context: context);
 
   @override
@@ -77,6 +82,9 @@ class _TwitterService implements TwitterService {
 
   @override
   final GeoService geoService;
+
+  @override
+  final TrendsService trendsService;
 
   @override
   final ComplianceService complianceService;